feat: remove ComponentEffect#bindChildren usage
Deprecates ComponentEffect#bindChildren. Remove all calls to it by using HasComponents#bindChildren.

Part of #23534
